rvv: fix the llu type differenct between 32/64 hosts

+++ riscv/insns/vnclip_vi.h
@@ -5,9 +5,9 @@ int64_t int_min = -(1 << (P.VU.vsew - 1));
 VI_VVXI_LOOP_NARROW
 ({
   int64_t result = vs2;
-  uint64_t unsigned_shift_amount = (uint64_t)(zimm5 & ((sew * 2) - 1));
+  unsigned unsigned_shift_amount = zimm5 & ((sew * 2) - 1);
   // The immediate form supports shift amounts up to 31 only.
-  unsigned_shift_amount = std::min(unsigned_shift_amount, 31ull);
+  unsigned_shift_amount = std::min(unsigned_shift_amount, 31u);
 
   // rounding
   INT_ROUNDING(result, xrm, unsigned_shift_amount);
